+++ nuxeo/auth/__init__.py
@@ -1,0 +1,7 @@
+# coding: utf-8
+from __future__ import unicode_literals
+
+from .portal_sso import PortalSSOAuth
+from .token import TokenAuth
+
+__all__ = ("PortalSSOAuth", "TokenAuth")